Adding more command line options.
This is good and now I understand what flexibility I have in command line options but now it is time
to write the database code up so that it can be used soon. I think that I will even need a
configuration file once I am done so that the user can specify what database they want to use and
other little options that they might want to pass in at any point.

The configuration should be read in first, then the command line arguments should be created, then
we should read the command from the command line and then we should setup the database if we need to
and then we should execute the actual command that needs to be executed.