Adding in the database schemas and some data.
You can see how the database is starting to form and it will be fantastic once it does. It is
quickly becoming something that will be able to do everything that I want of it. I just wish that
there was an easier way to create database schemas. Maybe that can be a future project.