Made the show command for htodo.
This is the first part that I will implement since it will only require reading from the database.
That way I can put the data that I want in the database and start testing what it is like to query
that data. This will make it much nicer in the future to work on the database schema because I will
have an implementation to work from. This is going to require clever work to make the filters and
the like come out correctly. Doubtless however the filters should translate nicely to some sql
commands that I can use.