Gave the show command basic functionality.
And ran accross some really odd ways that the HDBC library SqlValue's work that I will have to be
careful of in the future. I think that I will want to start writing the filter code and perpaps a
SQL generator. The more work that I can give to the database and save the Haskell program from
performing while still maintaining cross-database SQL syntax the better.