We have fixed more bugs and edge cases.
Mainly to do with what happens when things get initialized and the like.
All in all it is a much better effect in my opinion.

It is possible that I should change the code so that everything should be hidden by default.
Actually that is a fantastic idea. All of these todo lists will be hidden by default.