Adding in good indenting support.
Items should be indented correctly so that everything looks neat and the best way to do that is to
see what the maximum id in the database is and order the rest accordingly. I am also thinking of
introducing the concept of lists and only lists can have parents and children and lists contain
nothing but todo items.